fix: add exception

+++ AWS/EC2SnapShot.py
@@ -3,7 +3,9 @@ from pytz import timezone
 import boto3
 
 #########################
-REGION = "ap-northeast-2"
+REGION         = "ap-northeast-2"
+SNAPSHOT_COUNT = 5
+#스냅샷을 몇 개 남길 지 선택
 #########################
 
 def makeSnapshot():
@@ -22,32 +24,37 @@ def makeSnapshot():
         print(f'INSTANCE {instance.id} SNAPSHOT', end='')
         print('='*18)
         for volume in instance.volumes.all():
-            print(f'VOLUME : {volume.id} ..... ', end='')
-            snapshot = volume.create_snapshot(Description=f'{instance.id}/{volume.id}/{timestamp}')
-            print(f"SNAPSHOT ID : {snapshot.id}")
+            try:
+                print(f'VOLUME : {volume.id} ..... ', end='')
+                snapshot = volume.create_snapshot(Description=f'{instance.id}/{volume.id}/{timestamp}')
+                print(f"SNAPSHOT ID : {snapshot.id}")
+            except Exception as e:
+                print(f'ERROR : fail to make snapshot {volume.id} / {e}')
+                continue
 
 def clearSnapshot():
     account_id = boto3.client('sts').get_caller_identity().get('Account')
-    
     ec2 = boto3.client('ec2', region_name=REGION)
     response = ec2.describe_snapshots(OwnerIds=[account_id])
     snapshots = response["Snapshots"]
     snapshots.sort(key=lambda x: x["StartTime"])
-    print(f'현재 스냅샷 개수 : {len(snapshots)}\n'+'-'*25+'삭제할 스냅샷'+'-'*25)
-    snapshots = snapshots[:-2]
-    for snap in snapshots:
-        print(snap["Description"])
-    for snapshot in snapshots:
-        id = snapshot['SnapshotId']
-        try:
-            ec2.delete_snapshot(SnapshotId=id)
-            print(f'{id} 삭제 성공')
-        except Exception as e:
-            print(f'ERROR : fail to remove {id} / {e}')
-            continue
+    print(f'현재 스냅샷 개수 : {len(snapshots)}\n'+'-'*23+'스냅샷 삭제 시작'+'-'*23)
+    snapshots = snapshots[:-SNAPSHOT_COUNT]
+    if len(snapshots) != 0:
+        for snap in snapshots:
+            print(snap["Description"])
+            id = snap['SnapshotId']
+            try:
+                ec2.delete_snapshot(SnapshotId=id)
+                print(f'{id} 삭제 성공')
+            except Exception as e:
+                print(f'ERROR : fail to remove {id} / {e}')
+                continue
+    else:
+        print('삭제할 스냅샷 없음'.rjust(32))
 
 def lambda_handler(event, context):
-    # makeSnapshot()
+    makeSnapshot()
     clearSnapshot()
 
 lambda_handler('','')
